+++ app/api/terra/route.ts
@@ -1,0 +1,6 @@
+import { clerkClient } from '@clerk/clerk-sdk-node'
+
+export async function GET(request: Request) {
+  const clerk = clerkClient({ apiKey: process.env.CLERK_API_KEY })
+  return new Response('Hello, Terra!')
+}
