+++ app/patient/page.tsx
@@ -1,7 +1,30 @@
-export default async function () {
+import { Button } from '@/components/ui/button'
+import {
+  Dialog,
+  DialogContent,
+  DialogDescription,
+  DialogHeader,
+  DialogTitle,
+  DialogTrigger
+} from '@/components/ui/dialog'
+
+import { ProfileForm } from './components/form'
+
+export default function Page() {
   return (
-    <div>
-      <h1>PATIENT</h1>
-    </div>
+    <Dialog>
+      <DialogTrigger asChild>
+        <Button variant="outline">Edit Profile</Button>
+      </DialogTrigger>
+      <DialogContent className="sm:max-w-[425px]">
+        <DialogHeader>
+          <DialogTitle>Edit profile</DialogTitle>
+          <DialogDescription>
+            Make changes to your profile here. Click save when you're done.
+          </DialogDescription>
+        </DialogHeader>
+        <ProfileForm />
+      </DialogContent>
+    </Dialog>
   )
 }
